LPS-132797 Move to a new project
