+++ build.properties
@@ -166,6 +166,7 @@
     #    apps/frontend-token,\
     #    apps/frontend-view-state,\
     #    apps/gogo-shell,\
+    #    apps/google-places,\
     #    apps/headless,\
     #    apps/hello-velocity,\
     #    apps/iframe,\
