+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/build.gradle
@@ -34,9 +34,7 @@ dependencies {
 	compileOnly project(":apps:layout:layout-dynamic-data-mapping-form-field-type-api")
 	compileOnly project(":apps:layout:layout-type-controller:layout-type-controller-api")
 	compileOnly project(":apps:petra:petra-portlet-url-builder")
-	compileOnly project(":apps:portal-settings:portal-settings-api")
 	compileOnly project(":apps:portal:portal-instance-lifecycle-api")
-	compileOnly project(":apps:site:site-api")
 	compileOnly project(":apps:static:osgi:osgi-util")
 	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 	compileOnly project(":apps:static:portal-configuration:portal-configuration-persistence-api")
