+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-lang/src/main/resources/content/Language.properties
@@ -149,7 +149,6 @@ enable-template-creation=Enable Template Creation
 enable-template-creation-description=If checked, enable the creation of dynamic data mapping template.
 enable-x=Enable {0}
 end-parameter-name=End Parameter Name
-ensure-that-both-googles-places-api-and-maps-javascript-api-are-enabled-for-that-key=Ensure that both Google's Places API and Maps JavaScript API are enabled for that key.
 enter-a-default-value=Enter a default value.
 enter-a-default-value-that-is-submitted-if-no-other-value-is-entered=Enter a default value that is submitted if no other value is entered.
 enter-a-descriptive-field-label-that-guides-users-to-enter-the-information-you-want=Enter a descriptive field label that guides users to enter the information you want.
@@ -192,8 +191,6 @@ ftl[stands-for]=FreeMarker
 geolocation-field-type-description=Select a location from a map.
 geolocation-field-type-label=Geolocation
 google-hangouts=Google Hangouts
-google-places=Google Places
-google-places-api-key=Google Places API Key
 grid-field-type-description=Select options from a matrix.
 grid-field-type-label=Grid
 group-separators=Group Separators
@@ -366,7 +363,6 @@ select-object=Select Object
 separator-field-type-description=Add a separator to organize the content.
 separator-field-type-label=Separator
 separators=Separators
-set-the-google-places-api-key-that-is-used-for-this-set-of-pages=Set the Google Places API Key that is used for this set of pages.
 severity=Severity
 show-all=Show All
 show-as-a-switcher=Show as a Switcher
