+++ modules/apps/google-places/google-places-web/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Google Places Web
+Bundle-SymbolicName: com.liferay.google.places.web
+Bundle-Version: 1.0.0
+Web-ContextPath: /google-places-web
