+++ modules/apps/google-places/google-places-web/src/main/java/com/liferay/google/places/web/internal/site/settings/configuration/admin/display/GooglePlacesSiteSettingsConfigurationScreenContributor.java
@@ -12,15 +12,17 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.form.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.google.places.web.internal.site.settings.configuration.admin.display;
 
-import com.liferay.dynamic.data.mapping.constants.GooglePlacesWebKeys;
-import com.liferay.dynamic.data.mapping.util.GooglePlacesUtil;
+import com.liferay.google.places.web.internal.constants.GooglePlacesWebKeys;
+import com.liferay.google.places.web.internal.util.GooglePlacesUtil;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
+import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 
 import java.util.Locale;
 
@@ -34,23 +36,23 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Rodrigo Paulino
  */
-@Component(service = PortalSettingsConfigurationScreenContributor.class)
-public class GooglePlacesPortalSettingsConfigurationScreenContributor
-	implements PortalSettingsConfigurationScreenContributor {
+@Component(service = SiteSettingsConfigurationScreenContributor.class)
+public class GooglePlacesSiteSettingsConfigurationScreenContributor
+	implements SiteSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "third-party";
+		return "third-party-applications";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/portal_settings/google_places.jsp";
+		return "/site_settings/google_places.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "third-party-places";
+		return "third-party-applications-places";
 	}
 
 	@Override
@@ -58,13 +60,8 @@ public class GooglePlacesPortalSettingsConfigurationScreenContributor
 		return LanguageUtil.get(
 			ResourceBundleUtil.getBundle(
 				locale,
-				GooglePlacesPortalSettingsConfigurationScreenContributor.class),
+				GooglePlacesSiteSettingsConfigurationScreenContributor.class),
 			"google-places");
-	}
-
-	@Override
-	public String getSaveMVCActionCommandName() {
-		return "/portal_settings/edit_company";
 	}
 
 	@Override
@@ -77,22 +74,26 @@ public class GooglePlacesPortalSettingsConfigurationScreenContributor
 		HttpServletRequest httpServletRequest,
 		HttpServletResponse httpServletResponse) {
 
-		PortalSettingsConfigurationScreenContributor.super.setAttributes(
+		SiteSettingsConfigurationScreenContributor.super.setAttributes(
 			httpServletRequest, httpServletResponse);
 
 		ThemeDisplay themeDisplay =
 			(ThemeDisplay)httpServletRequest.getAttribute(
 				WebKeys.THEME_DISPLAY);
 
+		Group group = themeDisplay.getSiteGroup();
+
 		httpServletRequest.setAttribute(
 			GooglePlacesWebKeys.GOOGLE_PLACES_API_KEY,
 			GooglePlacesUtil.getGooglePlacesAPIKey(
-				themeDisplay.getCompanyId()));
+				themeDisplay.getCompanyId(), group.getGroupId(),
+				_groupLocalService));
 	}
 
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.dynamic.data.mapping.form.web)"
-	)
+	@Reference
+	private GroupLocalService _groupLocalService;
+
+	@Reference(target = "(osgi.web.symbolicname=com.liferay.google.places.web)")
 	private ServletContext _servletContext;
 
 }
