+++ modules/apps/google-places/google-places-web/src/main/java/com/liferay/google/places/web/internal/util/GooglePlacesUtil.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.util;
+package com.liferay.google.places.web.internal.util;
 
-import com.liferay.dynamic.data.mapping.constants.GooglePlacesWebKeys;
+import com.liferay.google.places.web.internal.constants.GooglePlacesWebKeys;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
