+++ modules/apps/google-places/google-places-web/src/main/resources/META-INF/resources/portal_settings/google_places.jsp
