+++ modules/apps/google-places/google-places-web/src/main/resources/META-INF/resources/site_settings/google_places.jsp
