+++ modules/apps/google-places/google-places-web/src/main/resources/content/Language.properties
@@ -1,0 +1,4 @@
+ensure-that-both-googles-places-api-and-maps-javascript-api-are-enabled-for-that-key=Ensure that both Google's Places API and Maps JavaScript API are enabled for that key.
+google-places=Google Places
+google-places-api-key=Google Places API Key
+set-the-google-places-api-key-that-is-used-for-this-set-of-pages=Set the Google Places API Key that is used for this set of pages.
