+++ modules/modules.properties
@@ -2659,6 +2659,7 @@ project.name[com.liferay.frontend.view.state.service]=:apps:frontend-view-state:
 project.name[com.liferay.gogo.shell.client]=:util:gogo-shell-client
 project.name[com.liferay.gogo.shell.web]=:apps:gogo-shell:gogo-shell-web
 project.name[com.liferay.google.maps.web]=:apps:archived:google-maps-web
+project.name[com.liferay.google.places.web]=:apps:google-places:google-places-web
 project.name[com.liferay.headless.admin.content.api]=:apps:headless:headless-admin-content:headless-admin-content-api
 project.name[com.liferay.headless.admin.content.client]=:apps:headless:headless-admin-content:headless-admin-content-client
 project.name[com.liferay.headless.admin.content.impl]=:apps:headless:headless-admin-content:headless-admin-content-impl
